Ai diagnostic assistant bf
- Fixed history fetching for agents
- Fixed history serialization
- Fixed history updation
- changed db model property type
- removed llm_report appendix
- increased timeout for response in api.py
- changed Probability threshold for CNN
- Improved agents instructions
- Added model fallback for agents

+++ Frontend/pages/dashboard.py
@@ -237,7 +237,7 @@ def open_create_modal():
                     "allergies": parse_list(allergies),
                     "medications": parse_list(medications),
                     "height_cm": int(height),
-                    "weight_kg": float(weight)
+                    "weight_kg": int(weight)
                 }
 
                 if create_patient(new_patient_data):
